refactor(app): consolidate admin routes into [locale] tree, split admin god-component (#934)
Removes the parallel non-i18n route tree so every page lives under
[locale], and breaks up the 642-line dashboard/admin/page.tsx monolith.

- Add [locale]/dashboard/admin/page.tsx: a thin overview composed of
  the existing platform-analytics cards/charts plus quick-link
  navigation to users/disputes/audit/categories/moderation, using the
  already-defined getPlatformAnalytics()/exportPlatformAnalyticsCsv()
  helpers from lib/api.ts instead of the monolith's hand-rolled fetch +
  CSV-link duplication.
- Diffed the legacy "users" tab against the standalone
  [locale]/dashboard/admin/users/page.tsx: the standalone page is a
  strict superset (adds status column + suspend/unsuspend/ban actions),
  so the legacy tab contributes nothing unique and was dropped.
  "disputes" and "audit" were never actually tabs in the legacy file —
  just links out to the pages that already exist under [locale].
- The legacy file also had "categories" and "moderation" tabs with no
  [locale] equivalent anywhere in the app. Rather than silently drop
  working admin capabilities, extracted them into new standalone pages
  ([locale]/dashboard/admin/categories, .../moderation) mirroring the
  users/disputes/audit pattern, so no functionality is lost.
- Deleted the non-locale dashboard/admin/{page,loading}.tsx monolith
  (and the now-empty app/dashboard directory) once parity was confirmed.
- Note: did not wire in components/Admin/AdminAnalyticsDashboard.tsx as
  the issue suggested — it's dead code that fetches from Next.js API
  routes (/api/analytics/admin/*) that don't exist in this app (only
  the Express backend has matching handlers at different paths). Wiring
  it in would have broken the overview page, so the working
  AdminCharts-based analytics were kept instead.
- Moved auth-callback and wallet/history under [locale] so they're no
  longer special-cased out of the locale tree. Fixed their two
  internal hardcoded non-locale redirects (sessionStorage
  oauth_redirect fallback and the OAuth-error "back to sign in" link)
  to be locale-aware via next-intl's useLocale(). The API's OAuth
  redirect (APP_URL/auth-callback?token=...) needed no change since
  next-intl's localePrefix:'always' middleware already redirects
  unprefixed hits to the default locale while preserving the query
  string.
- Added e2e/admin.spec.ts: Playwright smoke tests confirming all six
  admin routes and the two relocated routes are reachable under a
  locale prefix and render without crashing, following the existing
  curator.spec.ts "loads without crashing" / unauthenticated-redirect
  pattern (no prior admin e2e coverage existed).

Verification: eslint and an isolated `tsc --noEmit` pass are clean for
every file touched here. Full-repo `type-check`, `next build`, and
`next dev` are all pre-existing broken on main, independent of this
change (missing packages/types entry in pnpm-workspace.yaml blocks
install; a next-intl plugin/config path mismatch blocks the dev
server; ~90 tsc errors on main pre-date this branch and live entirely
in files this PR doesn't touch) — flagged in the PR description rather
than fixed here to keep this diff scoped to #934.

Closes #934
